pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "sphinx-revealjs"
description = "Sphinx extension with theme to generate Reveal.js presentation"
authors = [{name = "Kazuya Takei", email = "myself@attakei.net"}]
readme = "README.rst"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: Sphinx",
    "Framework :: Sphinx :: Extension",
    "Framework :: Sphinx :: Theme",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Multimedia :: Graphics :: Presentation",
    "Topic :: Software Development",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: HTML",
]
dynamic = ["version"]
dependencies = [
    "docutils",
    "Sphinx",
    "packaging",
]

[project.optional-dependencies]
screenshot = [
    "playwright",
]
test = [
    "BeautifulSoup4 >=4.9.3,<5",
    "pytest >=6.2.5,<7",
    "python-magic",
]
doc = [
    "sphinx-rtd-theme >=0.5.1,<0.6",
]
demo = [
    "sphinx-intl >=2.0.1,<3",
    "sphinxcontrib-gtagjs >=0.2.0,<0.3",
    "sphinxcontrib-budoux",
    "sphinxcontrib-oembed",
    "sphinxcontrib-sass",
]

[project.urls]
Home = "https://attakei.github.io/sphinx-revealjs"
Documentation = "https://sphinx-revealjs.readthedocs.io/"
Repository = "https://github.com/attakei/sphinx-revealjs"

[tool.flit.module]
name = "sphinx_revealjs"

[tool.flit.sdist]
include = [
    "demo",
    "doc",
    "sphinx_revealjs/themes/sphinx_revealjs/static/revealjs4/",
    "tests",
    "tools",
]
exclude = [
    "demo/_build",
    "doc/_build",
]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
